requirement.txt
# ===== Core dependencies =====
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
python-dateutil==2.8.2

# ===== Visualization =====
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# ===== Data analysis =====
statsmodels==0.14.0
prophet==1.1.5
scipy==1.10.1

# ===== Report generation =====
fpdf2==2.7.5
PyYAML==6.0
dashscope>=1.15.0  #Qwen API
scipy==1.10.1
DASHSCOPE_API_KEY = sk-275ca2c9f92e4855a220a829b2c073c4 #  API key


# ===== Other Tools =====
tqdm==4.65.0      
python-dotenv==1.0.0 
openpyxl==3.1.2 